requirements.txt
PySide6>=6.6
skia-python>=87.6
Pillow>=10.0
